analyzer-agent/requirements.txt
google-adk
google-genai
a2a-sdk
uvicorn
python-dotenv
pyyaml
httpx
langgraph
google-adk
langchain_google_genai